+++ src/index.js
@@ -7,7 +7,7 @@ import './index.css';
 import App from './components/app';
 
 import initializeStore from './redux/store';
-import history from './redux/reducer/history'
+import history from './redux/history'
 
 const initialState = {};
 const store = initializeStore(initialState, history)
